[fix!] change README.md to README.txt
Some platforms may not give the option to display a "README profile;"
for instance, I believe that Codeberg does not offer one.

Therefore, because I want consistency, it is changed to a text file from
a markdown file in order to have better compatibility with displaying in
the bio for platforms like Codeberg.

BREAKING CHANGE: Change display format greatly and breaks all markdown
formatting

Signed-off-by: Jason Scheffel <contact@jasonscheffel.com>