[fix!] change file back to markdown
This is because srht in their bio section kept messing up the formatting
with the text.

As for how to deal with Codeberg because they do not give a readme
profile thing. Since they only allow up to 250 chars for the bio, which
I just found out, I am just going to forward the user to a link to a
repository called 'jason-scheffel' on the Codeberg account.

Signed-off-by: Jason Scheffel <contact@jasonscheffel.com>